+++ Example/ManaKit/AppDelegate.swift
@@ -26,7 +26,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         
         // ... or test the database
         ManaKit.sharedInstance.setupResources()
-        
+
         return true
     }